+++ tarantool-spring-data/tarantool-spring-data-40/src/main/java/io/tarantool/spring/data40/config/package-info.java
@@ -1,0 +1,3 @@
+/** Config package. */
+@org.jspecify.annotations.NullMarked
+package io.tarantool.spring.data40.config;