+++ tarantool-spring-data/tarantool-spring-data-40/src/main/java/io/tarantool/spring/data40/core/annotation/package-info.java
@@ -1,0 +1,3 @@
+/** Core annotations of data mapping. */
+@org.jspecify.annotations.NullMarked
+package io.tarantool.spring.data40.core.annotation;