+++ tarantool-spring-data/tarantool-spring-data-40/src/main/java/io/tarantool/spring/data40/core/mapping/package-info.java
@@ -1,0 +1,3 @@
+/** Base package for the mapping subsystem. */
+@org.jspecify.annotations.NullMarked
+package io.tarantool.spring.data40.core.mapping;