+++ tarantool-spring-data/tarantool-spring-data-40/src/main/java/io/tarantool/spring/data40/core/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Core package for integrating <a href="https://www.tarantool.io/ru/">Tarantool</a> with Spring
+ * concepts.
+ */
+@org.jspecify.annotations.NullMarked
+package io.tarantool.spring.data40.core;