+++ tarantool-spring-data/tarantool-spring-data-40/src/main/java/io/tarantool/spring/data40/repository/package-info.java
@@ -1,0 +1,3 @@
+/** Repository package. */
+@org.jspecify.annotations.NullMarked
+package io.tarantool.spring.data40.repository;